Separate app and content logic
