+++ rsbuild.config.ts
@@ -30,6 +30,10 @@ export default function (env: any = {}, argv: Record<string, any> = {}) {
           import: './src/bandcamp/content/content.ts',
           html: false,
         },
+        'bandcamp.content.app': {
+          import: './src/bandcamp/content/app.ts',
+          html: false,
+        },
       },
     },
     resolve: {
