+++ src/bandcamp/page/music/musicFilter.css
@@ -1,56 +1,7 @@
-/* Masonry Grid Styles */
-#music-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
-  grid-gap: 15px;
-  grid-auto-rows: 10px;
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.music-grid-item {
-  margin: 0;
-  padding: 0;
-  break-inside: avoid;
-  display: inline-block;
-  width: 100%;
-  position: relative;
-}
-
-.music-grid-item a {
-  display: block;
-  text-decoration: none;
-  color: inherit;
-}
-
-.music-grid-item .art {
-  width: 100%;
-  position: relative;
-  overflow: hidden;
-  border-radius: 4px;
-}
-
-.music-grid-item .art img {
-  width: 100%;
-  height: auto;
-  display: block;
-}
-
-.music-grid-item .title {
-  margin: 8px 0 0 0;
-  font-size: 12px;
-  line-height: 1.3;
-  word-wrap: break-word;
-}
-
 /* Filter Search Box */
 .bcx-filter-container {
   margin-bottom: 20px;
   padding: 15px;
-  background: #f5f5f5;
-  border-radius: 8px;
-  border: 1px solid #ddd;
 }
 
 .bcx-filter-input {
@@ -73,7 +24,6 @@
   display: block;
   margin-bottom: 8px;
   font-weight: bold;
-  color: #333;
 }
 
 /* Hidden items during filtering */
